adding spring forces contribution
I created a simple template (in a repo called showcase)  to add the contributions locally and then upload the entire folder.  I used the Mandelbulb and Ulam spiral folders as my guide.  The index.json file needs some additions:  I wasn't sure what the url should be for the processing code  example and the image file for the folder is missing.